requirements.txt
argon2-cffi==20.1.0
asgiref==3.5.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bleach==3.3.0
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
entrypoints==0.3
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
oauthlib==3.1.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
pigeon-jupyter==0.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.2
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
tzdata==2022.1
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
